+++ photon-model-adapters/aws/src/main/java/com/vmware/photon/controller/model/adapters/awsadapter/enumeration/AWSEnumerationAndCreationAdapterService.java
@@ -556,6 +556,7 @@ public class AWSEnumerationAndCreationAdapterService extends StatelessService {
             computeState.endpointLink = this.context.request.endpointLink;
             computeState.descriptionLink = cd.documentSelfLink;
             computeState.type = ComputeType.VM_HOST;
+            computeState.environmentName = ComputeDescription.ENVIRONMENT_NAME_AWS;
 
             computeState.powerState = PowerState.ON;